9pfs: convert V9fsXattr::copied_len and ::len to uint32_t
Linux limits the size of an extended attribute value and the size of an
extended attribute namelist to 64k (see /usr/include/linux/limits.conf).

v9fs_xattrcreate() already ensures a guest cannot create a bigger
extended attribute, and v9fs_xattr_write() ensures that ::copied_len
cannot go beyond ::len.

When probing the list of extended attributes or reading an extended
attribute from the host filesystem, anything bigger than 64k denotes
a serious bug: we should abort() if this ever happens instead of
storing the value in ::len.

With all these sanity checks, it is enough to use an uint32_t for both
fields.

+++ hw/9pfs/9p.h
@@ -171,8 +171,8 @@ typedef struct V9fsConf
 
 typedef struct V9fsXattr
 {
-    uint64_t copied_len;
-    uint64_t len;
+    uint32_t copied_len;
+    uint32_t len;
     void *value;
     V9fsString name;
     int flags;
